+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#3306](https://github.com/bbatsov/rubocop/issues/3306): Add autocorrection for `Style/EachWithObject`. ([@owst][])
 * Add new `Style/TernaryParentheses` cop. ([@drenmi][])
+* [#3136](https://github.com/bbatsov/rubocop/issues/3136): Add config for `UselessAccessModifier` so it can be made aware of ActiveSupport's `concerning` and `class_methods` methods. ([@maxjacobson][])
 
 ### Bug fixes